renaming existing modal to merge-modal
