+++ app/src/app/app.component.spec.ts
@@ -28,7 +28,7 @@ import { MatIconModule } from '@angular/material/icon';
 import { MatDividerModule } from '@angular/material/divider';
 import { MatButtonModule } from '@angular/material/button';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { ModalModule } from './components/merge-modal/modal.module';
+import { ModalModule } from './components/merge-modal/merge-modal.module';
 import { queryElement } from 'src/shared/functions';
 import { DisplayResultsModule } from './components/display-results/display-results.module';
 import { YamlRenderModule } from './components/yaml-render/yaml-render.module';
