+++ app/src/app/app.component.ts
@@ -14,7 +14,7 @@
 
 import { Component, ViewChild, AfterViewInit } from '@angular/core';
 import { MatDialog } from '@angular/material/dialog';
-import { ModalComponent } from './components/merge-modal/modal.component';
+import { ModalComponent } from './components/merge-modal/merge-modal.component';
 import { OperationSet } from 'src/shared/interfaces';
 import { MediaMatcher } from '@angular/cdk/layout';
 import { Router } from '@angular/router';
