+++ app/src/app/components/merge-modal/confirm-operation/confirm-operation.component.scss
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 @import '../../../../shared/constants.scss';
-@import '../modal.component.scss';
+@import '../merge-modal.component.scss';
 
 .confirm-operation {
   &-container {
