+++ app/src/app/components/merge-modal/defaults-file-upload/defaults-file-upload.component.scss
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 @import '../../../../shared/constants.scss';
-@import '../modal.component.scss';
+@import '../merge-modal.component.scss';
 
 .defaults-file {
   &-upload-input {
