+++ app/src/app/components/merge-modal/merge-modal.component.scss
