+++ app/src/app/components/merge-modal/resolve-conflict/resolve-conflict.component.scss
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-@import '../modal.component.scss';
+@import '../merge-modal.component.scss';
 
 .conflicts {
   &-container {
