+++ app/src/app/components/merge-modal/spec-files-upload/spec-files-upload.component.scss
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 @import '../../../../shared/constants.scss';
-@import '../modal.component.scss';
+@import '../merge-modal.component.scss';
 
 .spec-files {
   &-upload-input {
