+++ app/src/app/components/merge-modal/spec-name-input/spec-name-input.component.scss
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 @import '../../../../shared/constants.scss';
-@import '../modal.component.scss';
+@import '../merge-modal.component.scss';
 
 .spec-name {
   &-form {
